feat: 增加xadmin.sh 脚本，用来控制服务端启停等操作

+++ utils/install_centos_docker.sh
@@ -2,21 +2,23 @@
 
 utils_dir=$(dirname "$(readlink -f "$0")")
 
-which docker && \
-yum remove -y docker \
-  docker-client \
-  docker-client-latest \
-  docker-common \
-  docker-latest \
-  docker-latest-logrotate \
-  docker-logrotate \
-  docker-engine
-
-which docker || (yum install -y yum-utils  \
-    && yum-config-manager --add-repo https://download.docker.com/linux/centos/docker-ce.repo  \
-    && sed -i 's+https://download.docker.com+https://mirrors.tuna.tsinghua.edu.cn/docker-ce+' /etc/yum.repos.d/docker-ce.repo  \
-    && yum install -y docker-ce docker-ce-cli containerd.io docker-buildx-plugin docker-compose-plugin  \
-    && systemctl restart docker)
+if which docker &>/dev/null ;then
+  yum remove -y docker \
+    docker-client \
+    docker-client-latest \
+    docker-common \
+    docker-latest \
+    docker-latest-logrotate \
+    docker-logrotate \
+    docker-engine
+fi
+if ! which docker &>/dev/null ;then
+  yum install -y yum-utils  \
+      && yum-config-manager --add-repo https://download.docker.com/linux/centos/docker-ce.repo  \
+      && sed -i 's+https://download.docker.com+https://mirrors.tuna.tsinghua.edu.cn/docker-ce+' /etc/yum.repos.d/docker-ce.repo  \
+      && yum install -y docker-ce docker-ce-cli containerd.io docker-buildx-plugin docker-compose-plugin  \
+      && systemctl restart docker
+fi
 
 project_parent_dir=$(dirname "$(dirname "${utils_dir}")")
